+++ webserver/src/main/java/cn/qkl/webserver/vo/riskAccount/NetworkAccountEdgeVO.java
@@ -5,6 +5,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @Author sunxiaen
  * @Date 2023/8/23 15:14
@@ -29,7 +31,7 @@ public class NetworkAccountEdgeVO {
     String to;
 
     @ApiModelProperty("交易金额")
-    Long txAmount;
+    BigDecimal txAmount;
 
     @ApiModelProperty("交易笔数")
     Long txNum;
